% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_corr.R
\name{compute_corr}
\alias{compute_corr}
\title{Computa correlacao tidy}
\usage{
compute_corr(data, var1, var2)
}
\arguments{
\item{data}{dados de entrada}

\item{var1}{nome da variavel 1}

\item{var2}{nome da variavel 2}
}
\value{
Uma tibble com a correlacao de  Pearson  e o p-valor
}
\description{
Computa correlacao tidy
}
\examples{
compute_corr(data = faithful, var1 = eruptions, var2 = waiting)
}
