% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cool_sum}
\alias{cool_sum}
\title{Add together two numbers.}
\usage{
cool_sum(left, right = 0)
}
\arguments{
\item{left}{A number.}

\item{right}{A number, optional.}
}
\value{
The sum of both numbers.
}
\description{
Adds two numbers together, in a cool way.
}
\details{
Not much else to say! This is just a really cool function.
}
\examples{
cool_sum(-3,5)
cool_sum(pi)
}
