% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{coolSum}
\alias{coolSum}
\title{Acrescenta dois números(Título da função)}
\usage{
coolSum(left, right = 0)
}
\arguments{
\item{left}{Um número.}

\item{right}{Um número opcional.}
}
\value{
A soma dos dois números
}
\description{
Acrescenta dois números, ou retorna o número único (Sumário da função
}
\details{
não há mais que se possa dizer! É uma função gira!
}
\examples{
coolSum(-3, 5)
coolSum(pi)
}
