% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{dados}
\alias{dados}
\title{Motor Trend Car Road Tests}
\format{32 linhas
\describe{
  \item{mpg}{miles p/ galon}
  \item{cyl}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
dados
}
\description{
The data was extracted from the 1974
Motor Trend US magazine, and comprises
fuel consumption and 10 aspects of automobile
design and performance for 32 automobiles
(1973–74 models).
}
\keyword{datasets}

