% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividir.R
\name{dividir}
\alias{dividir}
\title{Divide Dois Numeros}
\usage{
dividir(x, y)
}
\arguments{
\item{x}{Um numero}

\item{y}{Um numero}
}
\value{
A divisao de \code{x} e \code{y}
}
\description{
\code{dividir()} retorna a divisão entre \code{x} e \code{y}.
}
\examples{
# Divide 50/2 = 25
dividir(50, 2)
}
