% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello(i, j)
}
\arguments{
\item{i}{é o parametro que não serve pra nada}

\item{j}{outro parametro}
}
\description{
Imprime Hello World
}
\details{
Essa função faz isso
}
\examples{
hello()

}

