% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicar.R
\name{multiplicar}
\alias{multiplicar}
\title{Multiplica Dois Numeros}
\usage{
multiplicar(x, y)
}
\arguments{
\item{x}{Um numero}

\item{y}{Um numero}
}
\value{
A produto entre `x` e `y`
}
\description{
`multiplicar()` retorna o produto entre `x` e `y`.
}
\examples{
# Multiplica 5 * 3 = 15
multiplicar(5, 3)
}
