% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacote.R
\docType{package}
\name{pacote}
\alias{pacote}
\title{pacote: Um pacote como exemplo}
\description{
Este pacote possue algumas funções como exemplo em como fazer um pacote no R.
}
\section{coolSum}{
 soma fantástica
A função coolSum ...
}

