% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Subtrai Dois Numeros}
\usage{
subtrair(x, y)
}
\arguments{
\item{x}{Um numero}

\item{y}{Um numero}
}
\value{
A subtracao entre \code{x}  e \code{y}
}
\description{
\code{subtrair()} retorna a subtracao entre \code{x} e \code{y}.
}
\examples{
# Subtrair 1-1 = 0
subtrair(1, 1)
}
