% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_pocentagem.R
\name{Taxa_Porc}
\alias{Taxa_Porc}
\title{Criando proporções em porcentagens}
\usage{
Taxa_Porc(x, z)
}
\arguments{
\item{x, z}{Duas variáveis de interesse}
}
\value{
A porcentagem de cada caso correspondente entre as variáveis
}
\description{
Esta função cria uma proporção entre duas variáveis e transforma a proporção em porcentagem.
}
