% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variavel_quadratrica.R
\name{var_quad}
\alias{var_quad}
\title{variável_quadratrica}
\usage{
var_quad(df, variavel, tipo)
}
\arguments{
\item{df}{Um dataframe .}

\item{variavel}{Uma string indicando qual e a coluna de df a ser plotada}

\item{tipo}{O tipo do grafico a ser plotado}
}
\value{
Retorna um data frame com uma variável ao quadrado
}
\description{
variável_quadratrica
}
\examples{
var_quad(penguinson,"body_mass_g",integer)
}
