% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{coolSum}
\alias{coolSum}
\title{Acrescenta dois numeros(Titulo da funcao)}
\usage{
coolSum(left, right = 0)
}
\arguments{
\item{left}{Um numero.}

\item{right}{Um numero opcional.}
}
\value{
A soma dos dois numeros
}
\description{
Acrescenta dois numeros, ou retorna o numero unico (Sumario da funcao)
}
\details{
nao ha mais que se possa dizer! E uma funcao gira!
}
\examples{
coolSum(-3, 5)
coolSum(pi)
}
