% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dados}
\alias{dados}
\title{Dados arbitrários}
\format{
Um array com uma sequência aleatória (1-1000).
}
\source{
\url{https://www.r-project.org/}
}
\usage{
dados
}
\description{
Dados arbitrários como exmplo, obtidos com \code{sample(1000)}.
}
\keyword{datasets}
