% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicar.R
\name{multiplicar}
\alias{multiplicar}
\title{Multiplica Dois Numeros}
\usage{
multiplicar(x, y)
}
\arguments{
\item{x}{Um numero}

\item{y}{Um numero}
}
\value{
A produto entre \code{x} e \code{y}
}
\description{
\code{multiplicar()} retorna o produto entre \code{x} e \code{y}.
}
\examples{
# Multiplica 5 * 3 = 15
multiplicar(5, 3)
}
