% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{Soma Dois Numeros}
\usage{
somar(x, y)
}
\arguments{
\item{x}{Um numero}

\item{y}{Um numero}
}
\value{
A soma de `x` e `y`
}
\description{
`somar()` retorna a soma de `x` e `y`.
}
\examples{
# Soma 1+1 = 2
somar(1, 1)
}
